backend/requirements.txt
Flask
flask_cors
numpy
Pillow
python-dotenv
tensorflow
